Extract template-dx-integration
