+++ src/core/template-wrapper.ts
@@ -12,6 +12,8 @@ interface ITemplateWrapperProps {
     key: string;
 }
 
+type TemplateWrapperRenderer = () => TemplateWrapper;
+
 class TemplateWrapper extends React.PureComponent<ITemplateWrapperProps> {
     private readonly _removalListenerRef = React.createRef<HTMLElement>();
 
@@ -67,5 +69,6 @@ class TemplateWrapper extends React.PureComponent<ITemplateWrapperProps> {
 
 export {
     ITemplateWrapperProps,
-    TemplateWrapper
+    TemplateWrapper,
+    TemplateWrapperRenderer
 };
